feat: find and find_inode_id for directory entry lookup

+++ fs/src/fs.rs
@@ -75,11 +75,11 @@ impl FileSystem {
         Arc::new(Mutex::new(fs))
     }
 
-    pub fn get_disk_inode_pos(&self, inode_id: u32) -> (u32, u32) {
+    pub fn get_disk_inode_pos(&self, inode_id: u32) -> (u32, usize) {
         let inode_size = core::mem::size_of::<DiskInode>();
         let inode_per_block = (BLOCK_SIZE / inode_size) as u32;
         let inode_bid = self.inode_area_start + inode_id / inode_per_block;
-        let inode_offset = (inode_id % inode_per_block) * inode_size as u32;
+        let inode_offset = (inode_id % inode_per_block) as usize * inode_size;
         (inode_bid, inode_offset)
     }
